+++ cmake/thirdparty/BLTSetupMPI.cmake
@@ -1,6 +1,6 @@
 # Copyright (c) 2017-2023, Lawrence Livermore National Security, LLC and
 # other BLT Project Developers. See the top-level LICENSE file for details
-# 
+#
 # SPDX-License-Identifier: (BSD-3-Clause)
 
 ################################
@@ -32,14 +32,14 @@ set(_mpi_libraries )
 set(_mpi_link_flags )
 
 
-if(ENABLE_FIND_MPI)
+if(BLT_ENABLE_FIND_MPI)
     message(STATUS "FindMPI Enabled  (ENABLE_FIND_MPI == ON)")
 else()
     message(STATUS "FindMPI Disabled (ENABLE_FIND_MPI == OFF) ")
 endif()
 
 
-if (ENABLE_FIND_MPI)
+if (BLT_ENABLE_FIND_MPI)
     find_package(MPI REQUIRED)
 
     #-------------------
@@ -50,7 +50,7 @@ if (ENABLE_FIND_MPI)
     # Compile flags
     #-------------------
     set(_c_flag ${MPI_C_${_mpi_compile_flags_suffix}})
-    if (_c_flag AND ENABLE_CUDA)
+    if (_c_flag AND BLT_ENABLE_CUDA)
         list(APPEND _mpi_compile_flags
                     $<$<NOT:$<COMPILE_LANGUAGE:CUDA>>:${_c_flag}>
                     $<$<COMPILE_LANGUAGE:CUDA>:-Xcompiler=${_c_flag}>)
@@ -60,7 +60,7 @@ if (ENABLE_FIND_MPI)
 
     set(_cxx_flag ${MPI_CXX_${_mpi_compile_flags_suffix}})
     if (_cxx_flag AND NOT "${_c_flag}" STREQUAL "${_cxx_flag}")
-        if (ENABLE_CUDA)
+        if (BLT_ENABLE_CUDA)
             list(APPEND _mpi_compile_flags
             $<$<NOT:$<COMPILE_LANGUAGE:CUDA>>:${_cxx_flag}>
             $<$<COMPILE_LANGUAGE:CUDA>:-Xcompiler=${_cxx_flag}>)
@@ -69,7 +69,7 @@ if (ENABLE_FIND_MPI)
         endif()
     endif()
 
-    if (ENABLE_FORTRAN)
+    if (BLT_ENABLE_FORTRAN)
         set(_f_flag ${MPI_Fortran_${_mpi_compile_flags_suffix}})
         if (_f_flag AND NOT "${_c_flag}" STREQUAL "${_f_flag}")
             list(APPEND _mpi_compile_flags ${_f_flag})
@@ -84,7 +84,7 @@ if (ENABLE_FIND_MPI)
     #-------------------
     list(APPEND _mpi_includes ${MPI_C_${_mpi_includes_suffix}}
                               ${MPI_CXX_${_mpi_includes_suffix}})
-    if (ENABLE_FORTRAN)
+    if (BLT_ENABLE_FORTRAN)
         list(APPEND _mpi_includes ${MPI_Fortran_${_mpi_includes_suffix}})
     endif()
     blt_list_remove_duplicates(TO _mpi_includes)
@@ -96,7 +96,7 @@ if (ENABLE_FIND_MPI)
     if (NOT "${MPI_C_LINK_FLAGS}" STREQUAL "${MPI_CXX_LINK_FLAGS}")
         list(APPEND _mpi_link_flags ${MPI_CXX_LINK_FLAGS})
     endif()
-    if (ENABLE_FORTRAN)
+    if (BLT_ENABLE_FORTRAN)
         if ((NOT "${MPI_C_LINK_FLAGS}" STREQUAL "${MPI_Fortran_LINK_FLAGS}") AND
             (NOT "${MPI_CXX_LINK_FLAGS}" STREQUAL "${MPI_Fortran_LINK_FLAGS}"))
             list(APPEND _mpi_link_flags ${MPI_Fortran_LINK_FLAGS})
@@ -113,7 +113,7 @@ if (ENABLE_FIND_MPI)
         # Convert rpath flag if linking with CUDA
         string(REPLACE "-Wl,-rpath," "-Xlinker -rpath -Xlinker "
                        _mpi_link_flags "${_mpi_link_flags}")
-        # -pthread just doesn't work with nvcc                       
+        # -pthread just doesn't work with nvcc
         string(REPLACE "-pthread" " "
                        _mpi_link_flags "${_mpi_link_flags}")
     endif()
@@ -122,7 +122,7 @@ if (ENABLE_FIND_MPI)
     # Libraries
     #-------------------
     set(_mpi_libraries ${MPI_C_LIBRARIES} ${MPI_CXX_LIBRARIES})
-    if (ENABLE_FORTRAN)
+    if (BLT_ENABLE_FORTRAN)
         list(APPEND _mpi_libraries ${MPI_Fortran_LIBRARIES})
     endif()
     blt_list_remove_duplicates(TO _mpi_libraries)
@@ -166,14 +166,14 @@ endif()
 message(STATUS "MPI Num Proc Flag:    ${MPIEXEC_NUMPROC_FLAG}")
 message(STATUS "MPI Command Append:   ${BLT_MPI_COMMAND_APPEND}")
 
-if (ENABLE_FORTRAN)
+if (BLT_ENABLE_FORTRAN)
     # Determine if we should use fortran mpif.h header or fortran mpi module
     find_path(mpif_path
         NAMES "mpif.h"
         PATHS ${_mpi_includes}
         NO_DEFAULT_PATH
         )
-    
+
     if(mpif_path)
         set(MPI_Fortran_USE_MPIF ON CACHE PATH "")
         message(STATUS "Using MPI Fortran header: mpif.h")
@@ -183,10 +183,13 @@ if (ENABLE_FORTRAN)
     endif()
 endif()
 
+# Don't create the mpi target if it already exists.
 blt_import_library(NAME          mpi
                    INCLUDES      ${_mpi_includes}
                    TREAT_INCLUDES_AS_SYSTEM ON
                    LIBRARIES     ${_mpi_libraries}
                    COMPILE_FLAGS ${_mpi_compile_flags}
-                   LINK_FLAGS    ${_mpi_link_flags} 
+                   LINK_FLAGS    ${_mpi_link_flags}
                    EXPORTABLE    ${BLT_EXPORT_THIRDPARTY})
+
+add_library(blt::mpi ALIAS mpi)