+++ RELEASE-NOTES.md
@@ -16,9 +16,18 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
   the `clang-apply-replacements` executable in addition to the `CLANGTIDY_EXECUTABLE`.
   Also adds a corresponding `ENABLE_CLANGAPPLYREPLACEMENTS` CMake option.
   Note that the `clang_tidy_style` target is not added to the `style` target and must be run separately.
+- Added the `blt_install_tpl_setups` macro, which installs files to setup and create
+  targets for the third-party libraries OpenMP, MPI, CUDA, and HIP.  This macro is meant to 
+  replace `blt_export_tpl_targets` as the preferred way to setup third-party libraries with BLT.
+- Added `blt::`` namespaced aliases for BLT targets, `cuda`, `cuda_runtime`, `mpi`, and `openmp`.
+  These targets still exist but but will be deprecated in a future release. It is recommended that you
+  move to the new alias names, `blt::cuda`, `blt::cuda_runtime`, `blt::mpi`, and `blt::openmp`
 
 ### Changed
 - SetupHIP now searches for user-defined or environment variables before CMake paths to find the ROCM_PATH.
+
+### Fixed
+- Fixed infinite loop in `blt_find_target_dependencies`
 
 ## [Version 0.5.3] - Release date 2023-06-05